refactor(stop-subagent-enforcement): extract domain modules for maintainability
- Refactored 995 LOC monolithic file into 8 focused domain modules
- Created modular structure in scripts/hooks/stop-subagent-enforcement/:
  - config.js (128 LOC) - Configuration and requirements
  - time-utils.js (41 LOC) - Timezone normalization
  - bypass-handler.js (96 LOC) - Bypass file handling
  - post-completion-validator.js (128 LOC) - Post-completion validation
  - type-aware-validator.js (180 LOC) - Type-specific validation
  - bias-detector.js (178 LOC) - AI bias detection
  - sub-agent-validator.js (198 LOC) - Sub-agent validation
  - index.js (148 LOC) - Main orchestration
- Converted original file to re-export wrapper (57 LOC)
- Fixed validator to check REGRESSION instead of TESTING for refactor SDs
- Added Phase 4 refactoring orchestrator creation script
- All modules under 200 LOC target achieved

SD-LEO-REFAC-STOP-HOOK-004
